requirements.txt
openai>=1.12.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
tavily-python>=0.3.0
pydantic>=2.5.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
rich>=13.7.0
